경험치 증가량 추가

+++ Assets/Scripts/UI/ExpSlider.cs
@@ -13,7 +13,7 @@ public class ExpSlider : MonoBehaviour
     }
 
     void LateUpdate() {
-        slider.value = GameManager.Instance.Exp;
-        slider.maxValue = GameManager.Instance.MaxExp;
+        slider.value = (int)GameManager.Instance.Exp;
+        slider.maxValue = (int)GameManager.Instance.MaxExp;
     }
 }
